catalyst-trading-system/scripts/requirements_old.txt
# ============================================================================
# Catalyst Trading System - Python Requirements
# ============================================================================
# Name of Application: Catalyst Trading System
# Name of file: requirements.txt
# Version: 1.0.0
# Last Updated: 2025-11-18
# Purpose: Python package dependencies for all services
#
# REVISION HISTORY:
# v1.0.0 (2025-11-18) - Initial requirements
#   - Database connectivity packages
#   - Async support
#   - Environment management
#   - API framework
# ============================================================================

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# Environment Management
python-dotenv==1.0.0           # Load .env files

# Database - PostgreSQL
psycopg2-binary==2.9.9         # PostgreSQL adapter (with binary)
asyncpg==0.29.0                # Async PostgreSQL client
sqlalchemy==2.0.23             # SQL toolkit and ORM (optional)

# ============================================================================
# SERVICE DEPENDENCIES
# ============================================================================

# Web Framework (for services)
fastapi==0.104.1               # Modern web API framework
uvicorn[standard]==0.24.0      # ASGI server
pydantic==2.5.0                # Data validation

# Async Support
aiohttp==3.9.1                 # Async HTTP client/server
aiofiles==23.2.1               # Async file operations
asyncio==3.4.3                 # Async I/O (usually built-in)

# ============================================================================
# TRADING & MARKET DATA
# ============================================================================

# Market Data
alpaca-py==0.13.3              # Alpaca Markets API
yfinance==0.2.33               # Yahoo Finance data
pandas==2.1.4                  # Data manipulation
numpy==1.26.2                  # Numerical computing

# Technical Analysis
ta==0.11.0                     # Technical Analysis library
talib-binary==0.4.19           # TA-Lib (requires system library)

# ============================================================================
# UTILITIES
# ============================================================================

# Redis
redis==5.0.1                   # Redis client
aioredis==2.0.1               # Async Redis client

# HTTP & APIs
requests==2.31.0               # HTTP library
httpx==0.25.2                 # Modern HTTP client

# Date/Time
python-dateutil==2.8.2         # Date utilities
pytz==2023.3.post1            # Timezone support

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================

# Testing
pytest==7.4.3                  # Testing framework
pytest-asyncio==0.21.1         # Async test support
pytest-cov==4.1.0             # Coverage reporting

# Code Quality
black==23.12.0                 # Code formatter
flake8==6.1.0                 # Linting
mypy==1.7.1                   # Type checking

# Logging
loguru==0.7.2                 # Better logging
colorama==0.4.6               # Colored terminal output

# ============================================================================
# MONITORING & OBSERVABILITY
# ============================================================================

# Metrics
prometheus-client==0.19.0      # Prometheus metrics

# ============================================================================
# OPTIONAL ENHANCEMENTS
# ============================================================================

# Data Processing
polars==0.19.19               # Fast DataFrame library (alternative to pandas)

# Scheduling
apscheduler==3.10.4           # Advanced scheduler (for cron-like jobs)

# Configuration
pyyaml==6.0.1                 # YAML support
toml==0.10.2                  # TOML config files